chore: code cleanups + migrate to protocols

+++ terraflex/server/tf_state_lock_controller.py
@@ -8,19 +8,21 @@ from terraflex.server.base_state_lock_provider import (
     LockingError,
 )
 from terraflex.server.storage_provider_base import (
-    AbstractStorageProvider,
+    LockableStorageProviderProtocol,
+    StorageProviderProtocol,
     ItemKey,
+    WriteableStorageProviderProtocol,
 )
 from terraflex.server.transformation_base import (
-    AbstractTransformation,
+    TransformerProtocol,
 )
 
 
 class TFStateLockController(StateLockProviderProtocol):
     def __init__(
         self,
-        storage_driver: AbstractStorageProvider,
-        data_transformers: Iterable[AbstractTransformation],
+        storage_driver: StorageProviderProtocol,
+        data_transformers: Iterable[TransformerProtocol],
         state_file_storage_identifier: ItemKey,
     ):
         self.state_storage_identifier = state_file_storage_identifier
@@ -41,6 +43,9 @@ class TFStateLockController(StateLockProviderProtocol):
         return json.loads(content)
 
     async def put(self, lock_id: str, value: Data) -> None:
+        if not isinstance(self.storage_driver, WriteableStorageProviderProtocol):
+            raise NotImplementedError("This storage provider does not support writing")
+
         await self._check_lock(lock_id)
         # lock is locked by me
 
@@ -51,12 +56,18 @@ class TFStateLockController(StateLockProviderProtocol):
         self.storage_driver.put_file(self.state_storage_identifier, data)
 
     async def delete(self, lock_id: str) -> None:
+        if not isinstance(self.storage_driver, WriteableStorageProviderProtocol):
+            raise NotImplementedError("This storage provider does not support writing")
+
         await self._check_lock(lock_id)
         # lock is locked by me
 
         self.storage_driver.delete_file(self.state_storage_identifier)
 
     async def read_lock(self) -> LockBody | None:
+        if not isinstance(self.storage_driver, LockableStorageProviderProtocol):
+            raise NotImplementedError("This storage provider does not support writing")
+
         try:
             data = self.storage_driver.read_lock(self.state_storage_identifier)
 
@@ -66,6 +77,12 @@ class TFStateLockController(StateLockProviderProtocol):
         return LockBody.model_validate_json(data)
 
     async def _check_lock(self, lock_id: str) -> LockBody:
+        if not isinstance(self.storage_driver, LockableStorageProviderProtocol):
+            # This storage provider does not support locking
+            return LockBody(
+                ID="0000000000000000000", Operation="read", Who="me", Version="1", Created="2000-01-01T00:00:00Z"
+            )
+
         data = await self.read_lock()
         if data is None:
             raise LockingError(
@@ -82,7 +99,13 @@ class TFStateLockController(StateLockProviderProtocol):
         return data
 
     def lock(self, data: LockBody) -> None:
+        if not isinstance(self.storage_driver, LockableStorageProviderProtocol):
+            return
+
         self.storage_driver.acquire_lock(self.state_storage_identifier, data)
 
     def unlock(self) -> None:
+        if not isinstance(self.storage_driver, LockableStorageProviderProtocol):
+            return
+
         self.storage_driver.release_lock(self.state_storage_identifier)
